Bot Calling: Add incident call meeting. (#599)
Showcases using the APIs to create a scheduled meeting, installing the app to the meeting, inviting people, and sending a message to the chat.

Required a number of changes like removing the ability to support meeting events (as they require RSC and it's impossible to install an app with RSC using the APIs).
I've also remove the use of conversation caches and instead I'm using the adaptive card to include information about what call it is referring too.